+++ pialert/publishers/__init__.py
@@ -1,0 +1,8 @@
+""" Publishers for Pi.Alert """
+
+"""
+each publisher exposes:
+
+- check_config () returning True / False
+- send (message) returning True / Fasle 
+"""
